+++ packages/dara-core/js/components/index.tsx
@@ -6,6 +6,7 @@ export { default as TopBarFrame } from './top-bar-frame';
 export { default as DefaultFallback } from './fallback/default';
 export { default as RowFallback } from './fallback/row';
 export { default as CustomFallback } from './fallback/custom';
+export { default as Match } from './match';
 export { default as For } from './for/for';
 export { default as Link } from './link';
 export { default as Outlet } from './outlet';