+++ packages/dara-core/changelog.md
@@ -8,6 +8,7 @@ title: Changelog
 -  Added `is_first`, `is_last` and `index` properties to the `LoopVariable` class exposed as `Variable().list_item`. These "meta" properties can be useful inside the `For` component renderer
 -  Added support for parametrized routes in the `Link` component and `navigate` action by passing optional dynamic `params` kwarg
 -  Added `route_matches` property to the Router instance which lets users access the current route matches as a `Variable`. This can be useful for e.g. dynamic breadcrumb components
+-  Implemented `Match` component to allow conditional rendering based on a variable value without needing to use multiple `If` components
 
 ## 1.21.23